service-user/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
aiofiles==23.2.1
minio==7.1.17
pika==1.3.2
python-dotenv==1.0.0
httpx==0.25.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy[asyncio]==2.0.20
asyncpg==0.28.0
psycopg2-binary==2.9.7
alembic==1.12.0
email-validator==2.0.0
PyJWT==2.8.0
